% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_xy2TKA.R
\name{convert_xy2TKA}
\alias{convert_xy2TKA}
\title{Export xy-data to TKA}
\usage{
convert_xy2TKA(object, file = NULL, overwrite = FALSE)
}
\arguments{
\item{object}{\link{rxylib} (\strong{required}): xy data as imported by the function \link{read_xyData}. Optional
a file supported by the \code{rxylib}-package can be provided as input. Arguments can be provided as \link{list}.}

\item{file}{\link{character} (optional): optional file path or file name for the output to be written.
If only a path is provided the output file name is derived from the input file name. Argument
can be provided as \link{list}.}

\item{overwrite}{\link{logical} (with default): force overwriting of existing files if set to \code{TRUE}}
}
\value{
Returns a \link{list} of \link{matrix} objects or an output TKA-file.
}
\description{
Export data to the Toolkit file format (TKA) as exported by, e.g., by the software
Canberra Genie 2000.
}
\details{
\strong{Supported formats}
\itemize{
\item Canberra CNF
\item further formats on request ...
}
}
\section{Function version}{
 0.1.0
}

\examples{

##convert CNF data
convert_xy2TKA(
 object = system.file("extdata/ExampleSpectrum.CNF", package = "rxylib"))

\dontrun{
##export as file
convert_xy2TKA(
 object = system.file("extdata/ExampleSpectrum.CNF", package = "rxylib"),
 file = "~/Desktop/")

}

}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Université Bordeaux Montaigne (France)
}
\keyword{IO}
