% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_rxylib.R
\name{methods_rxylib}
\alias{methods_rxylib}
\alias{print.rxylib}
\alias{plot.rxylib}
\title{methods_rxylib}
\usage{
\method{print}{rxylib}(x, ...)

\method{plot}{rxylib}(x, block = NULL, ...)
}
\arguments{
\item{x}{(\strong{required}): input object}

\item{...}{further arguments that can be passed to the method}

\item{block}{\link{numeric} (with default): select block for plotting, e.g. \code{c(1:2)}.}
}
\description{
S3-methods support by the package \code{rxylib}. Listed functions can be passed directly into
S3 generics (e.g., \link{plot}, \link{print}) without reshaping the data.
}
