% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xyData.R
\name{read_xyData}
\alias{read_xyData}
\title{Import xy-Data for Supported Formats into R}
\usage{
read_xyData(file, verbose = TRUE)
}
\arguments{
\item{file}{\link{character} (\strong{required}): file to be imported}

\item{verbose}{\link{logical} (\emph{with default}): enables/disables verbose mode}
}
\description{
The function provides an access to the underlying \code{xylib} to import data for supported file formats
into R
}
\section{Function version}{
 0.1.0
}

\examples{

## still missing

}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Universite Bordeaux Montaigne (France)
}
\keyword{IO}
